Add DEBUG preprocessor flag
There is a bug (but it doesn't seem to change sqlite-queue behaviour).
This bug occurs in specific condtions so it is not urgent to fix it.
Eventhough calloc allocates memory and fills memory with zeros.
After read to allocated memory variable has bigger length (strlen) than, value specified in calloc,
and variable has greater length than output of malloc_usable_size which is crazy!!!

This error does not occur when sqlite-queue is running in valgrind